fix(History): fetch data before rendering the Panel to prevent flash
